Add macOS specific configuration for Yubikeys
To allow use of PKCS11-style (token-based) authentication with a
Yubikey. Specifically, macOS requires a custom keystore to be
whitelisted before it can be used with the default ssh-agent.

Spin up our own ssh-agent listening on a socket in ~/.ssh that has the
opensc PKCS11 store whitelisted.

Add helper additional helper functions.

Signed-off-by: Nick Travers <n.e.travers@gmail.com>